refactor(cli): split query expression into ast, scan, parser, and model
The query expression module had grown into a single 1000-line file
mixing the parsed-query types, string scanning, Cypher parsing, and
graph evaluation. It is now a thin facade over four focused submodules
(ast, scan, parser, model). Pure code move: the public surface and its
tests are unchanged.
